+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY13Z3K38TXMJ0PB4AWS92MC"
+        "01KY1VCC1HWP3KYAP80M5MF43S"
     )
     r"""The unique identifier of the data source"""