+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY13Z3JWD0P25DE4WKXT4S1X"
+        "01KY1VCC1AKJVHX86RD57HRE43"
     )
     r"""The unique identifier of the data source"""