+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-21T01:16:37.138Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-21T08:06:01.267Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")