+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY13Z3K70E4H66SM32NH8RB8"
+        "01KY1VCC1PT6XD86S0X43HNS9A"
     )
     r"""The unique identifier of the data source"""